move data validation/sanitization to front-end
